+++ CRM/GoCardlessUtils.php
@@ -313,7 +313,12 @@ class CRM_GoCardlessUtils {
         $params['installments'] = $installments;
       }
 
-      $result = static::completeRedirectFlowWithGoCardless($deets + $params);
+      $result = [];
+      // allow hook to amend params or override completeRedirectFlowWithGoCardless()
+      CRM_GoCardless_Hook::handleRedirectFlowWithGoCardless($deets + $params, $result);
+      if (empty($result)) {
+        $result = static::completeRedirectFlowWithGoCardless($deets + $params);
+      }
       // It's the subscription we're interested in.
       $subscription = $result['subscription'];
     }